@@ -16,19 +16,36 @@
         }
     ],
     "require": {
-        "php": ">=5.3"
+        "php": ">=8.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.8",
-        "squizlabs/php_codesniffer": "^2.6"
+        "squizlabs/php_codesniffer": "^2.6",
+        "rector/rector": "^0.18.8",
+        "pestphp/pest": "^2.24",
+        "pestphp/pest-plugin-drift": "^2.5",
+        "phpstan/phpstan": "^1.10",
+        "friendsofphp/php-cs-fixer": "^3.38"
     },
     "autoload": {
         "psr-4": {
-            "Sirius\\Html\\": "src/"
+            "Sirius\\Html\\": "src/",
+            "Tests\\": "tests/"
         }
     },
     "scripts": {
-        "cs": "phpcs --standard=PSR2 src/",
-        "cbf": "phpcbf --standard=PSR2 src/"
+        "stan": [
+            "php vendor/bin/phpstan analyse"
+        ],
+        "csfix": [
+            "php vendor/bin/php-cs-fixer fix src"
+        ],
+        "test": [
+            "php vendor/bin/pest --coverage --coverage-html=coverage"
+        ]
+    },
+    "config": {
+        "allow-plugins": {
+            "pestphp/pest-plugin": true
+        }
     }
 }
