@@ -1,9 +1,10 @@
 <?php
+
 namespace Sirius\Html\Tag;
 
 use Sirius\Html\Tag;
 
 class Button extends Tag
 {
-    protected $tag = 'button';
+    protected string $tag = 'button';
 }
