@@ -1,4 +1,5 @@
 <?php
+
 namespace Sirius\Html\Tag;
 
 class MultiSelect extends Select
@@ -8,21 +9,22 @@ class MultiSelect extends Select
      *
      * @return string
      */
-    protected function getOptionsString()
+    protected function getOptionsString(): string
     {
         $value   = $this->getValue();
         $options = '';
         if ($this->get('_first_option')) {
-            $options .= sprintf('<option value="">%s</option>', $this->get('_first_option'));
+            $options .= sprintf('<option value="">%s</option>', $this->get('_first_option')); // @phpstan-ignore-line
         }
-        $optionsList = $this->get('_options') ?: array();
+        /** @var array<int|string,string> $optionsList */
+        $optionsList = $this->get('_options') ?: [];
         foreach ($optionsList as $k => $v) {
             $selected = '';
             // be flexible, accept a non-array value
             if ((is_string($value) && $k == $value) || (is_array($value) && in_array($k, $value))) {
                 $selected = 'selected="selected"';
             }
-            $options .= sprintf('<option value="%s" %s>%s</option>', htmlentities($k, ENT_COMPAT), $selected, $v);
+            $options .= sprintf('<option value="%s" %s>%s</option>', htmlentities((string) $k, ENT_COMPAT), $selected, $v);
         }
 
         return $options;
@@ -30,8 +32,9 @@ class MultiSelect extends Select
 
     public function render()
     {
+        /** @var string $name */
         $name = $this->get('name');
-        if (substr($name, - 2) !== '[]') {
+        if (!str_ends_with($name, '[]')) {
             $this->set('name', $name . '[]');
         }
 
