@@ -1,37 +1,38 @@
 <?php
+
 namespace Sirius\Html\Tag;
 
 class Select extends Input
 {
-    protected $tag = 'select';
+    protected string $tag = 'select';
 
-    protected $isSelfClosing = false;
+    protected bool $isSelfClosing = false;
 
     /**
      * Generates the string with the list of the <OPTIONS> elements
-     *
-     * @return string
      */
-    protected function getOptionsString()
+    protected function getOptionsString(): string
     {
         $value   = $this->getValue();
         $options = '';
         if ($this->get('_first_option')) {
+            // @phpstan-ignore-next-line
             $options .= sprintf('<option value="">%s</option>', $this->get('_first_option'));
         }
-        $optionsList = $this->get('_options') ?: array();
+        /** @var array<string|int,string> $optionsList */
+        $optionsList = $this->get('_options') ?: [];
         foreach ($optionsList as $k => $v) {
             $selected = '';
             if ((is_string($value) && $k == $value) || (is_array($value) && in_array($k, $value))) {
                 $selected = 'selected="selected"';
             }
-            $options .= sprintf('<option value="%s" %s>%s</option>', htmlentities($k, ENT_COMPAT), $selected, $v);
+            $options .= sprintf('<option value="%s" %s>%s</option>', htmlentities((string) $k, ENT_COMPAT), $selected, $v);
         }
 
         return $options;
     }
 
-    public function getInnerHtml()
+    public function getInnerHtml(): string
     {
         return $this->getOptionsString();
     }
