@@ -1,18 +1,13 @@
 <?php
-namespace Sirius\Html\Tag;
 
-class HiddenTest extends \PHPUnit_Framework_TestCase
-{
+use \Sirius\Html\Tag\Hidden;
 
-    function setUp()
-    {
-        $this->input = new Hidden(array(
-            'name' => 'token'
-        ), '123');
-    }
+beforeEach(function () {
+    $this->input = new Hidden(array(
+        'name' => 'token'
+    ), '123');
+});
 
-    function testRender()
-    {
-        $this->assertEquals('<input name="token" type="hidden" value="123" />', (string) $this->input);
-    }
-}
+test('render', function () {
+    expect((string) $this->input)->toEqual('<input name="token" type="hidden" value="123" />');
+});
