@@ -1,20 +1,14 @@
 <?php
-namespace Sirius\Html\Tag;
 
-class RadioTest extends \PHPUnit_Framework_TestCase
-{
+use \Sirius\Html\Tag\Radio;
 
-    function setUp()
-    {
-        $this->input = new Radio(array(
-            'name'  => 'gender',
-            'value' => 'male'
-        ), 'male');
-    }
+beforeEach(function () {
+    $this->input = new Radio(array(
+        'name'  => 'gender',
+        'value' => 'male'
+    ), 'male');
+});
 
-    function testRender()
-    {
-        $this->assertEquals('<input checked="checked" name="gender" type="radio" value="male" />',
-            (string) $this->input);
-    }
-}
+test('render', function () {
+    expect((string) $this->input)->toEqual('<input checked="checked" name="gender" type="radio" value="male" />');
+});
