@@ -1,28 +1,23 @@
 <?php
-namespace Sirius\Html\Tag;
 
-class SelectTest extends \PHPUnit_Framework_TestCase
-{
+use \Sirius\Html\Tag\Select;
 
-    function setUp()
-    {
-        $this->input = new Select(array(
-            'name'          => 'answer',
-            'class'         => 'dropdown',
-            '_first_option' => '--select--',
-            '_options'      => array(
-                'yes'   => 'Yes',
-                'no'    => 'No',
-                'maybe' => 'Maybe'
-            )
-        ), 'maybe');
-    }
+beforeEach(function () {
+    $this->input = new Select(array(
+        'name'          => 'answer',
+        'class'         => 'dropdown',
+        '_first_option' => '--select--',
+        '_options'      => array(
+            'yes'   => 'Yes',
+            'no'    => 'No',
+            'maybe' => 'Maybe'
+        )
+    ), 'maybe');
+});
 
-    function testRender()
-    {
-        $result = (string) $this->input;
-        $this->assertTrue(strpos($result, '<select class="dropdown"') !== false);
-        $this->assertTrue(strpos($result, '<option value="maybe" selected="selected">Maybe</option>') !== false);
-        $this->assertTrue(strpos($result, '<option value="yes" >Yes</option>') !== false);
-    }
-}
+test('render', function () {
+    $result = (string) $this->input;
+    expect(strpos($result, '<select class="dropdown"') !== false)->toBeTrue();
+    expect(strpos($result, '<option value="maybe" selected="selected">Maybe</option>') !== false)->toBeTrue();
+    expect(strpos($result, '<option value="yes" >Yes</option>') !== false)->toBeTrue();
+});
