@@ -1,20 +1,14 @@
 <?php
-namespace Sirius\Html\Tag;
 
-class TextareaTest extends \PHPUnit_Framework_TestCase
-{
+use \Sirius\Html\Tag\Textarea;
 
-    function setUp()
-    {
-        $this->input = new Textarea(array(
-            'name' => 'comment',
-            'cols' => '30'
-        ), 'Sirius HTML rocks!');
-    }
+beforeEach(function () {
+    $this->input = new Textarea(array(
+        'name' => 'comment',
+        'cols' => '30'
+    ), 'Sirius HTML rocks!');
+});
 
-    function testRender()
-    {
-        $this->assertEquals('<textarea cols="30" name="comment">Sirius HTML rocks!</textarea>',
-            (string) $this->input);
-    }
-}
+test('render', function () {
+    expect((string) $this->input)->toEqual('<textarea cols="30" name="comment">Sirius HTML rocks!</textarea>');
+});
